style: add comments and some restructure
