+++ oid4vc-core/src/authorization_response.rs
@@ -2,6 +2,8 @@ use crate::openid4vc_extension::{Extension, ResponseHandle};
 use serde::{Deserialize, Serialize};
 use serde_with::skip_serializing_none;
 
+/// The [`AuthorizationResponse`] is a set of claims that are sent by a provider to a client. On top of some generic
+/// claims, it also contains a set of claims specific to an [`Extension`].
 #[skip_serializing_none]
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 pub struct AuthorizationResponse<E: Extension> {
