+++ oid4vp/src/oid4vp.rs
@@ -1,5 +1,5 @@
 use crate::authorization_request::{AuthorizationRequestBuilder, AuthorizationRequestParameters};
-use crate::oid4vp_params::Oid4vpParams;
+use crate::oid4vp_params::{serde_oid4vp_response, Oid4vpParams};
 use chrono::{Duration, Utc};
 pub use dif_presentation_exchange::{
     evaluate_input, ClaimFormatDesignation, InputDescriptor, InputDescriptorMappingObject, PathNested,
@@ -18,6 +18,7 @@ use serde::{Deserialize, Serialize};
 use serde_with::{DeserializeFromStr, SerializeDisplay};
 use std::sync::Arc;
 
+/// This is the [`RequestHandle`] for the [`OID4VP`] extension.
 #[derive(Debug, PartialEq)]
 pub struct RequestHandler {}
 impl RequestHandle for RequestHandler {
@@ -26,6 +27,7 @@ impl RequestHandle for RequestHandler {
     type Builder = AuthorizationRequestBuilder;
 }
 
+/// This is the [`ResponseHandle`] for the [`OID4VP`] extension.
 #[derive(Debug, PartialEq)]
 pub struct ResponseHandler {}
 impl ResponseHandle for ResponseHandler {
@@ -39,6 +41,7 @@ impl ResponseHandle for ResponseHandler {
 pub struct VpToken;
 serialize_unit_struct!("vp_token", VpToken);
 
+/// This is the [`Extension`] implementation for the [`OID4VP`] extension.
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 pub struct OID4VP;
 impl OpenID4VC for OID4VP {}
@@ -114,67 +117,6 @@ impl Extension for OID4VP {
 pub struct AuthorizationResponseParameters {
     #[serde(flatten, with = "serde_oid4vp_response")]
     pub oid4vp_parameters: Oid4vpParams,
-}
-
-pub mod serde_oid4vp_response {
-    use super::*;
-    use oid4vc_core::JsonValue;
-    use serde::{
-        de,
-        ser::{self, SerializeMap},
-    };
-
-    pub fn serialize<S>(oid4vp_response: &Oid4vpParams, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        match oid4vp_response {
-            Oid4vpParams::Jwt { response } => response.serialize(serializer),
-            Oid4vpParams::Params {
-                vp_token,
-                presentation_submission,
-            } => {
-                let mut map = serializer.serialize_map(Some(2))?;
-                map.serialize_entry("vp_token", vp_token)?;
-                map.serialize_entry(
-                    "presentation_submission",
-                    &serde_json::to_string(&presentation_submission).map_err(ser::Error::custom)?,
-                )?;
-                map.end()
-            }
-        }
-    }
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<Oid4vpParams, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        let oid4vp_response = JsonValue::deserialize(deserializer)?;
-        match oid4vp_response {
-            JsonValue::String(response) => Ok(Oid4vpParams::Jwt { response }),
-            JsonValue::Object(map) => {
-                let vp_token = map.get("vp_token").ok_or_else(|| {
-                    de::Error::custom(
-                        "`vp_token` parameter is required when using `presentation_submission` parameter.",
-                    )
-                })?;
-                let presentation_submission = map.get("presentation_submission").ok_or_else(|| {
-                    de::Error::custom(
-                        "`presentation_submission` parameter is required when using `vp_token` parameter.",
-                    )
-                })?;
-                let presentation_submission = presentation_submission
-                    .as_str()
-                    .ok_or_else(|| de::Error::custom("`presentation_submission` parameter must be a string."))?;
-                Ok(Oid4vpParams::Params {
-                    vp_token: serde_json::from_value(vp_token.clone()).map_err(de::Error::custom)?,
-                    presentation_submission: serde_json::from_str(presentation_submission)
-                        .map_err(de::Error::custom)?,
-                })
-            }
-            _ => Err(de::Error::custom("Invalid `oid4vp_response` parameter.")),
-        }
-    }
 }
 
 pub struct AuthorizationResponseInput {
